requirements.txt
Flask==2.3.4
python-dotenv==1.0.0
requests==2.31.0
pytest==8.2.2
